refactor: update icon and text color
- replace icon with scaled version
- set text color on heading
- tested in Lighthouse
