+++ src/layouts/SponsorsSection.jsx
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types'
 
-import Kite from '@/assets/images/kite.png'
+import Kite from '@/assets/images/icn-kite.png'
 
 const SponsorsSection = ({
   sponsorsData = [],
@@ -14,13 +14,14 @@ const SponsorsSection = ({
       className="flex flex-col justify-center p-8 sm:px-10 md:px-14 lg:px-16"
     >
       <div className="relative w-full pt-0">
-        <h2 className="w-full text-center font-russell text-4xl md:text-5xl lg:text-6xl">
+        <h2 className="w-full text-center font-russell text-4xl text-white md:text-5xl lg:text-6xl">
           {year && !isCurrentYear ? `${year} ` : ''}Sponsors
         </h2>
         <img
           src={Kite}
           alt="Sponsors"
           className="absolute right-2 top-6 h-12 sm:right-10 md:right-14 md:top-8 lg:right-16 lg:top-12 lg:h-16"
+          loading="lazy"
         />
       </div>
